enemy

+++ src/chain.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Default, Copy, Clone)]
 pub struct Chain {
-    value: Option<usize>,
+    value: Option<u32>,
     countdown: f64,
 }
 
@@ -15,11 +15,11 @@ impl Chain {
         self.countdown = 0.;
     }
 
-    pub fn get_value(&self) -> Option<usize> {
+    pub fn get_value(&self) -> Option<u32> {
         self.value
     }
 
-    pub fn update(&mut self, dt: f64) -> Option<usize> {
+    pub fn update(&mut self, dt: f64) -> Option<u32> {
         if self.countdown > 0. {
             self.countdown -= dt;
             if self.countdown <= 0. {
